requirements.txt
langchain>=0.2,<1.0
langchain-community>=0.2,<1.0
langchain-groq>=0.1,<1.0
chromadb>=0.4,<1.0
sentence-transformers>=2.2,<4.0
groq>=0.4,<1.0
streamlit>=1.30,<2.0
pandas>=2.0,<3.0
numpy>=1.24,<3.0
matplotlib>=3.7,<4.0
python-dotenv>=1.0,<2.0
pytest>=7.0,<9.0
